@@ -1,116 +1,143 @@
 {
   "name": "ng-alain",
-  "version": "0.6.5",
+  "version": "7.5.0",
   "description": "ng-zorro-antd admin panel front-end framework",
   "author": "cipchk",
-  "homepage": "https://cipchk.github.io/ng-alain/",
+  "homepage": "https://ng-alain.github.io/ng-alain/",
   "bugs": {
-    "url": "https://github.com/cipchk/ng-alain/issues"
-  },
-  "engines": {
-    "node": ">= 4.2.1",
-    "npm": ">= 3"
+    "url": "https://github.com/ng-alain/ng-alain/issues"
   },
   "license": "MIT",
+  "keywords": [
+    "delon",
+    "antd",
+    "ng-zorro-antd",
+    "angular",
+    "component",
+    "scaffold"
+  ],
   "scripts": {
-    "precommit": "npm run lint-staged",
     "ng": "ng",
-    "start": "ng serve -o",
-    "serve:hmr": "ng serve -o --hmr -e=hmr",
-    "build": "ng build --prod --build-optimizer",
+    "start": "npm run color-less && ng serve -o",
+    "hmr": "npm run color-less && ng serve -c=hmr",
+    "build": "npm run color-less && ng build --prod --build-optimizer",
     "analyze": "ng build --prod --build-optimizer --stats-json",
-    "lint": "run-s lint:ts lint:style",
-    "lint:ts": "ng lint",
-    "lint:style": "stylelint \"{src}/**/*.less\" --syntax less",
+    "lint": "npm run lint:ts && npm run lint:style",
+    "lint:ts": "tslint -p src/tsconfig.app.json -c tslint.json 'src/**/*.ts'",
+    "lint:style": "stylelint 'src/**/*.less' --syntax less",
     "lint-staged": "lint-staged",
+    "tslint-check": "tslint-config-prettier-check ./tslint.json",
     "e2e": "ng e2e",
-    "test": "ng test -sr -sm=false",
-    "test-coverage": "ng test -sr -cc -sm=false"
+    "test": "ng test --watch",
+    "test-coverage": "ng test --code-coverage --watch=false",
+    "color-less": "node scripts/color-less.js",
+    "icon": "ng g ng-alain:plugin icon"
   },
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@antv/data-set": "^0.8.3",
-    "@antv/g2": "^3.0.1",
-    "@antv/g2-plugin-slider": "^2.0.0",
-    "@delon/abc": "^0.6.5",
-    "@delon/acl": "^0.6.5",
-    "@delon/auth": "^0.6.5",
-    "@delon/theme": "^0.6.5",
-    "@delon/mock": "^0.6.5",
-    "@delon/cache": "^0.6.5",
-    "@ngx-translate/core": "^9.0.0",
-    "@ngx-translate/http-loader": "^2.0.0",
-    "angular-baidu-maps": "^1.0.1",
-    "angular-qq-maps": "^1.0.1",
-    "angular-split": "^1.0.0-rc.0",
-    "angular-tree-component": "^7.0.1",
-    "core-js": "^2.5.1",
-    "file-saver": "^1.3.3",
-    "font-awesome": "^4.7.0",
-    "moment": "^2.19.3",
-    "ng-clipboard-antd": "^2.0.1",
-    "ng-tree-antd": "^2.2.2",
-    "ng-zorro-antd": "^0.6.5",
-    "ng-zorro-antd-extra": "^1.1.6",
-    "ng2-dnd": "^5.0.0",
-    "ng2-img-cropper": "^0.9.0",
-    "ngx-color-picker": "^5.0.0",
-    "ngx-countdown": "^2.0.0",
-    "ngx-tinymce": "^1.0.1",
-    "ngx-ueditor": "^1.1.0",
-    "rxjs": "^5.5.5",
-    "screenfull": "^3.3.1",
-    "simple-line-icons": "^2.4.1",
-    "sweetalert2": "^7.0.0",
-    "weather-icons": "^1.3.2",
-    "zone.js": "^0.8.18"
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/router": "~7.2.0",
+    "core-js": "^2.5.4",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.26",
+    "@antv/data-set": "^0.10.2",
+    "@antv/g2": "^3.5.3",
+    "@antv/g2-plugin-slider": "^2.1.1",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "ajv": "^6.10.0",
+    "@delon/abc": "^7.5.0",
+    "@delon/acl": "^7.5.0",
+    "@delon/auth": "^7.5.0",
+    "@delon/cache": "^7.5.0",
+    "@delon/chart": "^7.5.0",
+    "@delon/form": "^7.5.0",
+    "@delon/mock": "^7.5.0",
+    "@delon/theme": "^7.5.0",
+    "@delon/util": "^7.5.0",
+    "file-saver": "^2.0.1",
+    "ng-zorro-antd": "^7.3.3",
+    "ngx-countdown": "^3.2.0",
+    "ngx-tinymce": "^7.0.0",
+    "ngx-ueditor": "^2.1.3",
+    "screenfull": "^4.2.0",
+    "qrious": "^4.0.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.2",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@angularclass/hmr": "^2.1.3",
-    "@angularclass/hmr-loader": "^3.0.4",
-    "@delon/cli": "^0.6.4",
-    "@types/jasmine": "~2.6.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "~7.3.9",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jszip": "^3.1.2",
-    "@types/node": "~6.0.60",
-    "codecov": "^3.0.0",
-    "codelyzer": "~4.0.1",
-    "jasmine-core": "~2.8.0",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-remap-istanbul": "^0.6.0",
-    "karma-sauce-launcher": "^1.2.0",
-    "lint-staged": "^5.0.0",
-    "npm-run-all": "^4.1.1",
-    "protractor": "~5.1.2",
-    "stylelint": "^8.2.0",
-    "stylelint-config-standard": "^17.0.0",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.6.0",
-    "webpack-bundle-analyzer": "^2.9.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2",
+    "@angularclass/hmr": "^2.1.3",
+    "@types/jszip": "^3.1.5",
+    "@types/mockjs": "^1.0.2",
+    "codecov": "^3.2.0",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-language-service": "^0.9.9",
+    "editorconfig-tools": "^0.1.1",
+    "husky": "^1.3.1",
+    "gh-pages": "^2.0.1",
+    "lint-staged": "^8.1.0",
+    "less-bundle-promise": "^1.0.7",
     "mockjs": "^1.0.1-beta3",
-    "@types/mockjs": "^1.0.0",
-    "xlsx": "^0.11.16"
+    "prettier": "^1.16.1",
+    "prettier-stylelint": "^0.4.2",
+    "stylelint": "^10.0.1",
+    "stylelint-config-css-modules": "^1.4.0",
+    "stylelint-config-prettier": "^5.1.0",
+    "stylelint-config-rational-order": "^0.1.2",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-declaration-block-no-ignored-properties": "^2.1.0",
+    "stylelint-order": "^3.0.0",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "antd-theme-generator": "^1.1.6",
+    "xlsx": "^0.14.2",
+    "@delon/testing": "^7.5.0",
+    "ng-alain": "^7.5.0",
+    "ng-alain-codelyzer": "^0.0.1"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint-staged"
+    }
   },
   "lint-staged": {
-    "src/**/*.ts": "lint:ts",
-    "src/**/*.less": "lint:style"
+    "*.{cmd,html,json,md,sh,txt,xml,yml}": [
+      "editorconfig-tools fix",
+      "git add"
+    ],
+    "*.ts": [
+      "npm run lint:ts",
+      "prettier --write",
+      "git add"
+    ],
+    "*.less": [
+      "stylelint 'src/**/*.less' --syntax less --fix",
+      "git add"
+    ],
+    "ignore": [
+      "src/assets/*"
+    ]
   }
 }
