@@ -1,55 +1,121 @@
-import { Injectable, Inject, Injector } from '@angular/core';
-import { Router } from '@angular/router';
-import { zhCN, enUS, NzLocaleService } from 'ng-zorro-antd';
+// 请参考：https://ng-alain.com/docs/i18n
+import { Injectable } from '@angular/core';
+import { Observable, BehaviorSubject } from 'rxjs';
+import { filter } from 'rxjs/operators';
+
+import { registerLocaleData } from '@angular/common';
+import ngZh from '@angular/common/locales/zh';
+import ngEn from '@angular/common/locales/en';
+import ngZhTw from '@angular/common/locales/zh-Hant';
+
+import { en_US, zh_CN, zh_TW, NzI18nService } from 'ng-zorro-antd';
+import * as df_en from 'date-fns/locale/en';
+import * as df_zh_cn from 'date-fns/locale/zh_cn';
+import * as df_zh_tw from 'date-fns/locale/zh_tw';
 import { TranslateService } from '@ngx-translate/core';
-import { SettingsService, AlainI18NService } from '@delon/theme';
-import { Observable } from 'rxjs/Observable';
+import {
+  SettingsService,
+  AlainI18NService,
+  DelonLocaleService,
+  en_US as delonEnUS,
+  zh_CN as delonZhCn,
+  zh_TW as delonZhTw,
+} from '@delon/theme';
 
-@Injectable()
+interface LangData {
+  text: string;
+  ng: any;
+  zorro: any;
+  dateFns: any;
+  delon: any;
+  abbr: string;
+}
+
+const DEFAULT = 'zh-CN';
+const LANGS: { [key: string]: LangData } = {
+  'zh-CN': {
+    text: '简体中文',
+    ng: ngZh,
+    zorro: zh_CN,
+    dateFns: df_zh_cn,
+    delon: delonZhCn,
+    abbr: '🇨🇳',
+  },
+  'zh-TW': {
+    text: '繁体中文',
+    ng: ngZhTw,
+    zorro: zh_TW,
+    dateFns: df_zh_tw,
+    delon: delonZhTw,
+    abbr: '🇭🇰',
+  },
+  'en-US': {
+    text: 'English',
+    ng: ngEn,
+    zorro: en_US,
+    dateFns: df_en,
+    delon: delonEnUS,
+    abbr: '🇬🇧',
+  },
+};
+
+@Injectable({ providedIn: 'root' })
 export class I18NService implements AlainI18NService {
+  private _default = DEFAULT;
+  private change$ = new BehaviorSubject<string | null>(null);
 
-    private _default = 'zh-CN';
+  private _langs = Object.keys(LANGS).map(code => {
+    const item = LANGS[code];
+    return { code, text: item.text, abbr: item.abbr };
+  });
 
-    private _langs = [
-        { code: 'en', text: 'English' },
-        { code: 'zh-CN', text: '中文' }
-    ];
+  constructor(
+    settings: SettingsService,
+    private nzI18nService: NzI18nService,
+    private delonLocaleService: DelonLocaleService,
+    private translate: TranslateService,
+  ) {
+    const defaultLan = settings.layout.lang || translate.getBrowserLang();
+    // `@ngx-translate/core` 预先知道支持哪些语言
+    const lans = this._langs.map(item => item.code);
+    translate.addLangs(lans);
 
-    constructor(
-        settings: SettingsService,
-        private nzLocalService: NzLocaleService,
-        private translate: TranslateService,
-        private injector: Injector
-    ) {
-        this._default = settings.layout.lang || translate.getBrowserLang();
-        const lans = this._langs.map(item => item.code);
-        if (!lans.includes(this._default)) {
-            this._default = lans[0];
-        }
-        translate.addLangs(lans);
-    }
+    this._default = lans.includes(defaultLan) ? defaultLan : lans[0];
+    this.updateLangData(this._default);
+  }
 
-    use(lang: string = null, firstLoad = true): Observable<any> {
-        lang = lang || this.translate.getDefaultLang();
-        this.nzLocalService.setLocale(lang === 'en' ? enUS : zhCN);
-        // need reload router because of ng-zorro-antd local system
-        if (!firstLoad) this.injector.get(Router).navigate([ '/' ]);
-        return this.translate.use(lang);
-    }
-    /** 获取语言列表 */
-    getLangs() {
-        return this._langs;
-    }
-    /** 翻译 */
-    fanyi(key: string) {
-        return this.translate.instant(key);
-    }
-    /** 默认语言 */
-    get defaultLang() {
-        return this._default;
-    }
-    /** 当前语言 */
-    get currentLang() {
-        return this.translate.currentLang || this.translate.getDefaultLang() || this._default;
-    }
+  private updateLangData(lang: string) {
+    const item = LANGS[lang];
+    registerLocaleData(item.ng);
+    this.nzI18nService.setLocale(item.zorro);
+    (window as any).__locale__ = item.dateFns;
+    this.delonLocaleService.setLocale(item.delon);
+  }
+
+  get change(): Observable<string> {
+    return this.change$.asObservable().pipe(filter(w => w != null)) as Observable<string>;
+  }
+
+  use(lang: string): void {
+    lang = lang || this.translate.getDefaultLang();
+    if (this.currentLang === lang) return;
+    this.updateLangData(lang);
+    this.translate.use(lang).subscribe(() => this.change$.next(lang));
+  }
+  /** 获取语言列表 */
+  getLangs() {
+    return this._langs;
+  }
+  /** 翻译 */
+  fanyi(key: string, interpolateParams?: Object) {
+    return this.translate.instant(key, interpolateParams);
+  }
+  /** 默认语言 */
+  get defaultLang() {
+    return this._default;
+  }
+  /** 当前语言 */
+  get currentLang() {
+    return this.translate.currentLang || this.translate.getDefaultLang() || this._default;
+  }
 }
