@@ -1,22 +1,32 @@
-import { Component, HostListener } from '@angular/core';
+import { Component, HostListener, ChangeDetectionStrategy } from '@angular/core';
 import * as screenfull from 'screenfull';
 
 @Component({
-    selector: 'header-fullscreen',
-    template: `
-    <i class="anticon anticon-{{status ? 'shrink' : 'arrows-alt'}}"></i>
-    {{(status ? 'fullscreen-exit' : 'fullscreen') | translate }}
-    `
+  selector: 'header-fullscreen',
+  template: `
+    <i nz-icon [nzType]="status ? 'fullscreen-exit' : 'fullscreen'"></i>
+    {{ (status ? 'menu.fullscreen.exit' : 'menu.fullscreen') | translate }}
+  `,
+  host: {
+    '[class.d-block]': 'true',
+  },
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class HeaderFullScreenComponent {
+  status = false;
+  private get sf(): screenfull.Screenfull {
+    return screenfull as screenfull.Screenfull;
+  }
 
-    status = false;
+  @HostListener('window:resize')
+  _resize() {
+    this.status = this.sf.isFullscreen;
+  }
 
-    @HostListener('click')
-    _click() {
-        if (screenfull.enabled) {
-            screenfull.toggle();
-        }
-        this.status = !screenfull.isFullscreen;
+  @HostListener('click')
+  _click() {
+    if (this.sf.enabled) {
+      this.sf.toggle();
     }
+  }
 }
