@@ -1,49 +1,51 @@
-import { Component, HostBinding, ViewChild, Input, OnInit, ElementRef, AfterViewInit } from '@angular/core';
+import { Component, HostBinding, Input, ElementRef, AfterViewInit, ChangeDetectionStrategy } from '@angular/core';
 
 @Component({
-    selector: 'header-search',
-    template: `
-    <nz-input nzPlaceHolder='{{ "top-search-ph" | translate}}' [(ngModel)]="q"
-        (nzFocus)="qFocus()" (nzBlur)="qBlur()">
-        <ng-template #prefix>
-            <i class="anticon anticon-search"></i>
-        </ng-template>
-    </nz-input>
-    `
+  selector: 'header-search',
+  template: `
+    <nz-input-group [nzAddOnBeforeIcon]="focus ? 'arrow-down' : 'search'">
+      <input
+        nz-input
+        [(ngModel)]="q"
+        (focus)="qFocus()"
+        (blur)="qBlur()"
+        [placeholder]="'menu.search.placeholder' | translate"
+      />
+    </nz-input-group>
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class HeaderSearchComponent implements AfterViewInit {
+  q: string;
 
-    q: string;
+  qIpt: HTMLInputElement;
 
-    qIpt: HTMLInputElement;
+  @HostBinding('class.alain-default__search-focus')
+  focus = false;
 
-    @HostBinding('class.header-search__focus')
-    focus = false;
+  @HostBinding('class.alain-default__search-toggled')
+  searchToggled = false;
 
-    @HostBinding('class.header-search__toggled')
-    searchToggled = false;
+  @Input()
+  set toggleChange(value: boolean) {
+    if (typeof value === 'undefined') return;
+    this.searchToggled = true;
+    this.focus = true;
+    setTimeout(() => this.qIpt.focus(), 300);
+  }
 
-    @Input()
-    set toggleChange(value: boolean) {
-        if (typeof value === 'undefined') return;
-        console.log('toggleChange', value);
-        this.searchToggled = true;
-        this.focus = true;
-        setTimeout(() => this.qIpt.focus(), 300);
-    }
+  constructor(private el: ElementRef) {}
 
-    constructor(private el: ElementRef) {}
+  ngAfterViewInit() {
+    this.qIpt = (this.el.nativeElement as HTMLElement).querySelector('.ant-input') as HTMLInputElement;
+  }
 
-    ngAfterViewInit() {
-        this.qIpt = (this.el.nativeElement as HTMLElement).querySelector('.ant-input') as HTMLInputElement;
-    }
+  qFocus() {
+    this.focus = true;
+  }
 
-    qFocus() {
-        this.focus = true;
-    }
-
-    qBlur() {
-        this.focus = false;
-        this.searchToggled = false;
-    }
+  qBlur() {
+    this.focus = false;
+    this.searchToggled = false;
+  }
 }
