@@ -1,28 +1,28 @@
-import { Component, HostListener } from '@angular/core';
+import { Component, HostListener, ChangeDetectionStrategy } from '@angular/core';
 import { NzModalService, NzMessageService } from 'ng-zorro-antd';
 
 @Component({
-    selector: 'header-storage',
-    template: `
-    <i class="anticon anticon-tool"></i>
-    {{ 'clear-local-storage' | translate}}`
+  selector: 'header-storage',
+  template: `
+    <i nz-icon nzType="tool"></i>
+    {{ 'menu.clear.local.storage' | translate }}
+  `,
+  host: {
+    '[class.d-block]': 'true',
+  },
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class HeaderStorageComponent {
+  constructor(private modalSrv: NzModalService, private messageSrv: NzMessageService) {}
 
-    constructor(
-        private confirmServ: NzModalService,
-        private messageServ: NzMessageService
-    ) {
-    }
-
-    @HostListener('click')
-    _click() {
-        this.confirmServ.confirm({
-            title: 'Make sure clear all local storage?',
-            onOk: () => {
-                localStorage.clear();
-                this.messageServ.success('Clear Finished!');
-            }
-        });
-    }
+  @HostListener('click')
+  _click() {
+    this.modalSrv.confirm({
+      nzTitle: 'Make sure clear all local storage?',
+      nzOnOk: () => {
+        localStorage.clear();
+        this.messageSrv.success('Clear Finished!');
+      },
+    });
+  }
 }
