@@ -1,12 +1,11 @@
-import { Component } from '@angular/core';
-import { NzMessageService } from 'ng-zorro-antd';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { SettingsService } from '@delon/theme';
 
 @Component({
-  selector   : 'app-sidebar',
-  templateUrl: './sidebar.component.html'
+  selector: 'layout-sidebar',
+  templateUrl: './sidebar.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class SidebarComponent {
-    constructor(public settings: SettingsService, public msgSrv: NzMessageService) {
-    }
+  constructor(public settings: SettingsService) {}
 }
