@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-    selector: 'layout-fullscreen',
-    templateUrl: './fullscreen.component.html'
+  selector: 'layout-fullscreen',
+  templateUrl: './fullscreen.component.html',
+  host: {
+    '[class.alain-fullscreen]': 'true',
+  },
 })
-export class LayoutFullScreenComponent {
-
-}
+export class LayoutFullScreenComponent {}
