@@ -1,79 +1,115 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, ChangeDetectorRef } from '@angular/core';
 import { NzMessageService } from 'ng-zorro-antd';
-import { getTimeDistance, yuan } from '@delon/abc';
+import { STColumn } from '@delon/abc';
+import { getTimeDistance } from '@delon/util';
 import { _HttpClient } from '@delon/theme';
+import { I18NService } from '@core';
+import { yuan } from '@shared';
 
 @Component({
-    selector: 'app-dashboard-analysis',
-    templateUrl: './analysis.component.html',
-    styleUrls: ['./analysis.component.less']
+  selector: 'app-dashboard-analysis',
+  templateUrl: './analysis.component.html',
+  styleUrls: ['./analysis.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DashboardAnalysisComponent implements OnInit {
-    data: any = {
-        salesData: [],
-        offlineData: []
-    };
-    loading = true;
-    q: any = {
-        start: null,
-        end: null
-    };
-    rankingListData: any[] = Array(7).fill({}).map((item, i) => {
-        return {
-            title: `工专路 ${i} 号店`,
-            total: 323234
-        };
+  data: any = {};
+  loading = true;
+  date_range: Date[] = [];
+  rankingListData: any[] = Array(7)
+    .fill({})
+    .map((item, i) => {
+      return {
+        title: this.i18n.fanyi('app.analysis.test', { no: i }),
+        total: 323234,
+      };
     });
+  titleMap = {
+    y1: this.i18n.fanyi('app.analysis.traffic'),
+    y2: this.i18n.fanyi('app.analysis.payments'),
+  };
+  searchColumn: STColumn[] = [
+    { title: '排名', i18n: 'app.analysis.table.rank', index: 'index' },
+    {
+      title: '搜索关键词',
+      i18n: 'app.analysis.table.search-keyword',
+      index: 'keyword',
+      click: (item: any) => this.msg.success(item.keyword),
+    },
+    {
+      type: 'number',
+      title: '用户数',
+      i18n: 'app.analysis.table.users',
+      index: 'count',
+      sorter: (a, b) => a.count - b.count,
+    },
+    {
+      type: 'number',
+      title: '周涨幅',
+      i18n: 'app.analysis.table.weekly-range',
+      index: 'range',
+      render: 'range',
+      sorter: (a, b) => a.range - b.range,
+    },
+  ];
 
-    constructor(private http: _HttpClient, public msg: NzMessageService) {}
+  constructor(
+    private http: _HttpClient,
+    public msg: NzMessageService,
+    private i18n: I18NService,
+    private cdr: ChangeDetectorRef,
+  ) {}
 
-    ngOnInit() {
-        this.http.get('/chart').subscribe((res: any) => {
-            res.offlineData.forEach((item: any) => {
-                item.chart = Object.assign([], res.offlineChartData);
-            });
-            this.data = res;
-            this.loading = false;
-            this.changeSaleType();
-        });
+  ngOnInit() {
+    this.http.get('/chart').subscribe((res: any) => {
+      res.offlineData.forEach((item: any, idx: number) => {
+        item.show = idx === 0;
+        item.chart = Object.assign([], res.offlineChartData);
+      });
+      this.data = res;
+      this.loading = false;
+      this.changeSaleType();
+    });
+  }
+
+  setDate(type: any) {
+    this.date_range = getTimeDistance(type);
+    setTimeout(() => this.cdr.detectChanges());
+  }
+
+  salesType = 'all';
+  salesPieData: any;
+  salesTotal = 0;
+  changeSaleType() {
+    this.salesPieData =
+      this.salesType === 'all'
+        ? this.data.salesTypeData
+        : this.salesType === 'online'
+        ? this.data.salesTypeDataOnline
+        : this.data.salesTypeDataOffline;
+    if (this.salesPieData) {
+      this.salesTotal = this.salesPieData.reduce((pre, now) => now.y + pre, 0);
     }
+    this.cdr.detectChanges();
+  }
 
-    setDate(type: any) {
-        const rank = getTimeDistance(type);
-        this.q.start = rank[0];
-        this.q.end = rank[1];
-    }
+  handlePieValueFormat(value: any) {
+    return yuan(value);
+  }
 
-    sort(sortName, sortValue) {
-        this.data.searchData = [
-            ...(<any[]>this.data.searchData).sort((a, b) => {
-                if (a[ sortName ] > b[ sortName ]) {
-                    return (sortValue === 'ascend') ? 1 : -1;
-                } else if (a[ sortName ] < b[ sortName ]) {
-                    return (sortValue === 'ascend') ? -1 : 1;
-                } else {
-                    return 0;
-                }
-            })
-        ];
+  saleTabs: any[] = [{ key: 'sales', show: true }, { key: 'visits' }];
+  salesChange(idx: number) {
+    if (this.saleTabs[idx].show !== true) {
+      this.saleTabs[idx].show = true;
+      this.cdr.detectChanges();
     }
+  }
 
-    salesType = 'all';
-    salesPieData: any;
-    salesTotal = 0;
-    changeSaleType() {
-        this.salesPieData = this.salesType === 'all' ? this.data.salesTypeData : (
-            this.salesType === 'online' ? this.data.salesTypeDataOnline : this.data.salesTypeDataOffline
-        );
-        if (this.salesPieData) this.salesTotal = this.salesPieData.reduce((pre, now) => now.y + pre, 0);
+  offlineIdx = 0;
+  offlineChange(idx: number) {
+    if (this.data.offlineData[idx].show !== true) {
+      this.data.offlineData[idx].show = true;
+      this.cdr.detectChanges();
     }
-
-    handlePieValueFormat(value: any) {
-        return yuan(value);
-    }
-
-    _activeTab = 0;
-    _tabChange(value: any) {
-        console.log('tab', this._activeTab, value);
-    }
+  }
 }
