@@ -2,12 +2,10 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { RelationComponent } from './relation/relation.component';
 
-const routes: Routes = [
-    { path: 'relation', component: RelationComponent }
-];
+const routes: Routes = [{ path: 'relation', component: RelationComponent }];
 
 @NgModule({
-  imports: [ RouterModule.forChild(routes) ],
-  exports: [ RouterModule ]
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
 })
-export class DataVRoutingModule { }
+export class DataVRoutingModule {}
