@@ -1,7 +1,14 @@
 import { Component } from '@angular/core';
+import { NzModalService } from 'ng-zorro-antd';
 
 @Component({
-    selector: 'exception-403',
-    template: `<exception type="403" style="min-height: 500px; height: 80%;"></exception>`
+  selector: 'exception-403',
+  template: `
+    <exception type="403" style="min-height: 500px; height: 80%;"></exception>
+  `,
 })
-export class Exception403Component {}
+export class Exception403Component {
+  constructor(modalSrv: NzModalService) {
+    modalSrv.closeAll();
+  }
+}
