@@ -1,7 +1,14 @@
 import { Component } from '@angular/core';
+import { NzModalService } from 'ng-zorro-antd';
 
 @Component({
-    selector: 'exception-404',
-    template: `<exception type="404" style="min-height: 500px; height: 80%;"></exception>`
+  selector: 'exception-404',
+  template: `
+    <exception type="404" style="min-height: 500px; height: 80%;"></exception>
+  `,
 })
-export class Exception404Component {}
+export class Exception404Component {
+  constructor(modalSrv: NzModalService) {
+    modalSrv.closeAll();
+  }
+}
