@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { SharedModule } from '@shared/shared.module';
+import { SharedModule } from '@shared';
 
 import { ExtrasRoutingModule } from './extras-routing.module';
 
@@ -9,19 +8,12 @@ import { ExtrasSettingsComponent } from './settings/settings.component';
 import { ExtrasPoiComponent } from './poi/poi.component';
 import { ExtrasPoiEditComponent } from './poi/edit/edit.component';
 
-const COMPONENTS_NOROUNT = [ ExtrasPoiEditComponent ];
+const COMPONENTS = [HelpCenterComponent, ExtrasSettingsComponent, ExtrasPoiComponent];
+const COMPONENTS_NOROUNT = [ExtrasPoiEditComponent];
 
 @NgModule({
-    imports: [
-        SharedModule,
-        ExtrasRoutingModule
-    ],
-    declarations: [
-        HelpCenterComponent,
-        ExtrasSettingsComponent,
-        ExtrasPoiComponent,
-        ...COMPONENTS_NOROUNT
-    ],
-    entryComponents: COMPONENTS_NOROUNT
+  imports: [SharedModule, ExtrasRoutingModule],
+  declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
+  entryComponents: COMPONENTS_NOROUNT,
 })
-export class ExtrasModule { }
+export class ExtrasModule {}
