@@ -1,9 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-    selector: 'pro-result-fail',
-    templateUrl: './fail.component.html'
+  selector: 'app-result-fail',
+  templateUrl: './fail.component.html',
 })
-export class ProResultFailComponent {
-
-}
+export class ProResultFailComponent {}
