@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 import { NzMessageService } from 'ng-zorro-antd';
 
 @Component({
-    selector: 'pro-result-success',
-    templateUrl: './success.component.html'
+  selector: 'app-result-success',
+  templateUrl: './success.component.html',
 })
 export class ProResultSuccessComponent {
-    constructor(public msg: NzMessageService) {}
+  constructor(public msg: NzMessageService) {}
 }
