@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { SimpleGuard } from '@delon/auth';
 import { environment } from '@env/environment';
 // layout
 import { LayoutDefaultComponent } from '../layout/default/default.component';
@@ -16,63 +17,80 @@ import { UserRegisterComponent } from './passport/register/register.component';
 import { UserRegisterResultComponent } from './passport/register-result/register-result.component';
 // single pages
 import { CallbackComponent } from './callback/callback.component';
-import { Exception403Component } from './exception/403.component';
-import { Exception404Component } from './exception/404.component';
-import { Exception500Component } from './exception/500.component';
+import { UserLockComponent } from './passport/lock/lock.component';
 
 const routes: Routes = [
-    {
-        path: '',
-        component: LayoutDefaultComponent,
-        children: [
-            { path: '', redirectTo: 'dashboard/v1', pathMatch: 'full' },
-            { path: 'dashboard', redirectTo: 'dashboard/v1', pathMatch: 'full' },
-            { path: 'dashboard/v1', component: DashboardV1Component, data: { translate: 'dashboard_v1' } },
-            { path: 'dashboard/analysis', component: DashboardAnalysisComponent, data: { translate: 'dashboard_analysis' } },
-            { path: 'dashboard/monitor', component: DashboardMonitorComponent, data: { translate: 'dashboard_monitor' } },
-            { path: 'dashboard/workplace', component: DashboardWorkplaceComponent, data: { translate: 'dashboard_workplace' } },
-            { path: 'widgets', loadChildren: './widgets/widgets.module#WidgetsModule' },
-            { path: 'elements', loadChildren: './elements/elements.module#ElementsModule' },
-            { path: 'other', loadChildren: './other/other.module#OtherModule' },
-            { path: 'forms', loadChildren: './forms/forms.module#FormsModule' },
-            { path: 'editor', loadChildren: './editor/editor.module#EditorModule' },
-            { path: 'charts', loadChildren: './charts/charts.module#ChartsModule' },
-            { path: 'tables', loadChildren: './tables/tables.module#TablesModule' },
-            { path: 'maps', loadChildren: './maps/maps.module#MapsModule' },
-            { path: 'pages', loadChildren: './pages/pages.module#PagesModule' },
-            { path: 'logics', loadChildren: './logics/logics.module#LogicsModule' },
-            { path: 'extras', loadChildren: './extras/extras.module#ExtrasModule' },
-            { path: 'pro', loadChildren: './pro/pro.module#ProModule' }
-        ]
-    },
-    // 全屏布局
-    {
-        path: 'data-v',
-        component: LayoutFullScreenComponent,
-        children: [
-            { path: '', loadChildren: './data-v/data-v.module#DataVModule' }
-        ]
-    },
-    // passport
-    {
-        path: 'passport',
-        component: LayoutPassportComponent,
-        children: [
-            { path: 'login', component: UserLoginComponent },
-            { path: 'register', component: UserRegisterComponent },
-            { path: 'register-result', component: UserRegisterResultComponent }
-        ]
-    },
-    // 单页不包裹Layout
-    { path: 'callback/:type', component: CallbackComponent },
-    { path: '403', component: Exception403Component },
-    { path: '404', component: Exception404Component },
-    { path: '500', component: Exception500Component },
-    { path: '**', redirectTo: 'dashboard' }
+  {
+    path: '',
+    component: LayoutDefaultComponent,
+    canActivate: [SimpleGuard],
+    canActivateChild: [SimpleGuard],
+    children: [
+      { path: '', redirectTo: 'dashboard/v1', pathMatch: 'full' },
+      { path: 'dashboard', redirectTo: 'dashboard/v1', pathMatch: 'full' },
+      { path: 'dashboard/v1', component: DashboardV1Component },
+      { path: 'dashboard/analysis', component: DashboardAnalysisComponent },
+      { path: 'dashboard/monitor', component: DashboardMonitorComponent },
+      { path: 'dashboard/workplace', component: DashboardWorkplaceComponent },
+      {
+        path: 'widgets',
+        loadChildren: './widgets/widgets.module#WidgetsModule',
+      },
+      { path: 'style', loadChildren: './style/style.module#StyleModule' },
+      { path: 'delon', loadChildren: './delon/delon.module#DelonModule' },
+      { path: 'extras', loadChildren: './extras/extras.module#ExtrasModule' },
+      { path: 'pro', loadChildren: './pro/pro.module#ProModule' },
+      // Exception
+      { path: 'exception', loadChildren: './exception/exception.module#ExceptionModule' },
+    ],
+  },
+  // 全屏布局
+  {
+    path: 'data-v',
+    component: LayoutFullScreenComponent,
+    children: [{ path: '', loadChildren: './data-v/data-v.module#DataVModule' }],
+  },
+  // passport
+  {
+    path: 'passport',
+    component: LayoutPassportComponent,
+    children: [
+      {
+        path: 'login',
+        component: UserLoginComponent,
+        data: { title: '登录', titleI18n: 'app.login.login' },
+      },
+      {
+        path: 'register',
+        component: UserRegisterComponent,
+        data: { title: '注册', titleI18n: 'app.register.register' },
+      },
+      {
+        path: 'register-result',
+        component: UserRegisterResultComponent,
+        data: { title: '注册结果', titleI18n: 'app.register.register' },
+      },
+      {
+        path: 'lock',
+        component: UserLockComponent,
+        data: { title: '锁屏', titleI18n: 'app.lock' },
+      },
+    ],
+  },
+  // 单页不包裹Layout
+  { path: 'callback/:type', component: CallbackComponent },
+  { path: '**', redirectTo: 'exception/404' },
 ];
 
 @NgModule({
-    imports: [RouterModule.forRoot(routes, { useHash: environment.useHash })],
-    exports: [RouterModule]
-  })
-export class RouteRoutingModule { }
+  imports: [
+    RouterModule.forRoot(routes, {
+      useHash: environment.useHash,
+      // NOTICE: If you use `reuse-tab` component and turn on keepingScroll you can set to `disabled`
+      // Pls refer to https://ng-alain.com/components/reuse-tab
+      scrollPositionRestoration: 'top',
+    }),
+  ],
+  exports: [RouterModule],
+})
+export class RouteRoutingModule {}
