@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 
-import { SharedModule } from '@shared/shared.module';
+import { SharedModule } from '@shared';
 import { RouteRoutingModule } from './routes-routing.module';
 // dashboard pages
 import { DashboardV1Component } from './dashboard/v1/v1.component';
@@ -12,28 +12,27 @@ import { UserLoginComponent } from './passport/login/login.component';
 import { UserRegisterComponent } from './passport/register/register.component';
 import { UserRegisterResultComponent } from './passport/register-result/register-result.component';
 // single pages
+import { UserLockComponent } from './passport/lock/lock.component';
 import { CallbackComponent } from './callback/callback.component';
-import { Exception403Component } from './exception/403.component';
-import { Exception404Component } from './exception/404.component';
-import { Exception500Component } from './exception/500.component';
+
+const COMPONENTS = [
+  DashboardV1Component,
+  DashboardAnalysisComponent,
+  DashboardMonitorComponent,
+  DashboardWorkplaceComponent,
+  // passport pages
+  UserLoginComponent,
+  UserRegisterComponent,
+  UserRegisterResultComponent,
+  // single pages
+  UserLockComponent,
+  CallbackComponent,
+];
+const COMPONENTS_NOROUNT = [];
 
 @NgModule({
-    imports: [ SharedModule, RouteRoutingModule ],
-    declarations: [
-        DashboardV1Component,
-        DashboardAnalysisComponent,
-        DashboardMonitorComponent,
-        DashboardWorkplaceComponent,
-        // passport pages
-        UserLoginComponent,
-        UserRegisterComponent,
-        UserRegisterResultComponent,
-        // single pages
-        CallbackComponent,
-        Exception403Component,
-        Exception404Component,
-        Exception500Component
-    ]
+  imports: [SharedModule, RouteRoutingModule],
+  declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
+  entryComponents: COMPONENTS_NOROUNT,
 })
-
 export class RoutesModule {}
