@@ -3,12 +3,10 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { WidgetsComponent } from './widgets/widgets.component';
 
-const routes: Routes = [
-    { path: '', component: WidgetsComponent }
-];
+const routes: Routes = [{ path: '', component: WidgetsComponent }];
 
 @NgModule({
-  imports: [ RouterModule.forChild(routes) ],
-  exports: [ RouterModule ]
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
 })
-export class WidgetsRoutingModule { }
+export class WidgetsRoutingModule {}
