@@ -1,15 +1,17 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { SharedModule } from '@shared/shared.module';
+import { SharedModule } from '@shared';
 
 import { WidgetsRoutingModule } from './widgets-routing.module';
 
 import { WidgetsComponent } from './widgets/widgets.component';
 
+const COMPONENTS = [WidgetsComponent];
+
+const COMPONENTS_NOROUNT = [];
+
 @NgModule({
-  imports: [ SharedModule, WidgetsRoutingModule ],
-  declarations: [
-    WidgetsComponent
-  ]
+  imports: [SharedModule, WidgetsRoutingModule],
+  declarations: [...COMPONENTS, ...COMPONENTS_NOROUNT],
+  entryComponents: COMPONENTS_NOROUNT,
 })
-export class WidgetsModule { }
+export class WidgetsModule {}
