@@ -3,58 +3,64 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 // delon
-import { NgZorroAntdExtraModule } from 'ng-zorro-antd-extra';
 import { AlainThemeModule } from '@delon/theme';
-import { AlainACLModule } from '@delon/acl';
-import { ZORROMODULES, ABCMODULES } from '../delon.module';
+import { DelonABCModule } from '@delon/abc';
+import { DelonChartModule } from '@delon/chart';
+import { DelonACLModule } from '@delon/acl';
+import { DelonFormModule } from '@delon/form';
 // i18n
 import { TranslateModule } from '@ngx-translate/core';
 
-// region: third libs
+// #region third libs
+import { NgZorroAntdModule } from 'ng-zorro-antd';
 import { CountdownModule } from 'ngx-countdown';
-const THIRDMODULES = [ CountdownModule ];
-// endregion
+import { UEditorModule } from 'ngx-ueditor';
+import { NgxTinymceModule } from 'ngx-tinymce';
 
-// region: your componets & directives
+const THIRDMODULES = [NgZorroAntdModule, CountdownModule, UEditorModule, NgxTinymceModule];
+// #endregion
+
+// #region your componets & directives
 const COMPONENTS = [];
 const DIRECTIVES = [];
-// endregion
+// #endregion
 
 @NgModule({
-    imports: [
-        CommonModule,
-        FormsModule,
-        RouterModule,
-        ReactiveFormsModule,
-        ...ZORROMODULES,
-        NgZorroAntdExtraModule,
-        AlainThemeModule.forChild(),
-        ...ABCMODULES,
-        AlainACLModule,
-        // third libs
-        ...THIRDMODULES
-    ],
-    declarations: [
-        // your components
-        ...COMPONENTS,
-        ...DIRECTIVES
-    ],
-    exports: [
-        CommonModule,
-        FormsModule,
-        ReactiveFormsModule,
-        RouterModule,
-        ...ZORROMODULES,
-        NgZorroAntdExtraModule,
-        AlainThemeModule,
-        ...ABCMODULES,
-        // i18n
-        TranslateModule,
-        // third libs
-        ...THIRDMODULES,
-        // your components
-        ...COMPONENTS,
-        ...DIRECTIVES
-    ]
+  imports: [
+    CommonModule,
+    FormsModule,
+    RouterModule,
+    ReactiveFormsModule,
+    AlainThemeModule.forChild(),
+    DelonABCModule,
+    DelonChartModule,
+    DelonACLModule,
+    DelonFormModule,
+    // third libs
+    ...THIRDMODULES,
+  ],
+  declarations: [
+    // your components
+    ...COMPONENTS,
+    ...DIRECTIVES,
+  ],
+  exports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    RouterModule,
+    AlainThemeModule,
+    DelonABCModule,
+    DelonChartModule,
+    DelonACLModule,
+    DelonFormModule,
+    // i18n
+    TranslateModule,
+    // third libs
+    ...THIRDMODULES,
+    // your components
+    ...COMPONENTS,
+    ...DIRECTIVES,
+  ],
 })
-export class SharedModule { }
+export class SharedModule {}
