@@ -1,7 +1,6 @@
 export const environment = {
-    chore: false,
-    SERVER_URL: `./`,
-    production: true,
-    hmr: false,
-    useHash: true
+  SERVER_URL: `./`,
+  production: true,
+  useHash: true,
+  hmr: false,
 };
