@@ -1,18 +1,8 @@
-/* SystemJS module definition */
-declare var module: NodeModule;
-interface NodeModule {
-  id: string;
-}
-
-interface WebpackRequire {
-  <T>(path: string): T;
-  (paths: string[], callback: (...modules: any[]) => void): void;
-  ensure: (paths: string[], callback: (require: <T>(path: string) => T) => void) => void;
-}
-interface NodeRequire extends WebpackRequire {}
-declare var require: NodeRequire;
+// # 3rd Party Library
+// If the library doesn't have typings available at `@types/`,
+// you can still use it by manually adding typings for it
 
 // G2
 declare var G2: any;
+declare var DataSet: any;
 declare var Slider: any;
-declare var Cloud: any;
